requirements.txt
scrapy 
ebooklib
requests
bs4
Pillow
googey

